Beta Version (so far)
Since we have no spritework, I've made a dot that moves around with the
PlayerMovement script. This dot should be replaced with a proper sprite
later.

I've also imported the PathedMovement and NextPoint scripts I wrote for
the TowerDefense game from last year. I've made a few changes so that
guards will wait before moving to the next point.

Be sure to look at all of the scripts and don't forget that even though
this is a 2D game, some 3D may be necessary. Improve them or add new
ones if you can. Guards cannot see at the moment and are harmless.